cargo clippy

+++ sfinder-lib/src/common/tetfu/field/array_colored_field.rs
@@ -24,7 +24,7 @@ impl ColoredField for ArrayColoredField {
         } else {
             let mut new_field = vec![[ColorType::Empty; FIELD_WIDTH as usize]; max_height as usize];
             for i in 0..max_height as usize {
-                new_field[i] = self.field[i as usize];
+                new_field[i] = self.field[i];
             }
             new_field
         };
